/fp:fast coverage
